requirements.txt
streamlitpandasrequests